TextInput case rename pt 1
