+++ jsapp/js/components/common/DebouncedTextInput.tsx
@@ -1,7 +1,7 @@
 import { useDebouncedCallback } from '@mantine/hooks'
 import React, { useCallback, useEffect, useRef, useState } from 'react'
-import TextInput from './textInput'
-import type { TextInputProps } from './textInput'
+import TextInput from './TextInputxxx'
+import type { TextInputProps } from './TextInputxxx'
 
 export interface DebouncedTextInputProps extends Omit<TextInputProps, 'value' | 'onChange'> {
   value?: string
